make parsing of ∀/∃ more intuitive Previously, the body of a ∀/∃ was parsed to be any expression.
This had some surprising results, such as parsing
  ∀xPx ∧ ∀yQy
as
  ∀x(Px ∧ ∀yQy)
instead of
  (∀xPx) ∧ (∀yQy)

This commit fixes this by assuming binary operators to be outside of the body of a ∀/∃, unless the body is surrounded by parenthesis.

More technically, we switch from parsing the body of a ∀/∃ as any expression to parsing it as a non-binary-operator expression, which includes the case of being parenthesized.